More realistic sinusoidal wave
Beginning to understand why this was hard to implement before, and it
has to do with how we're drawing the points. This currently does a
fair amount of curve compressing when the crest of the wave is hitting
the edge of the window, but I still think it looks fairly decent.