Render a different gradient for each hour of the day
Gradients cribbed from dribble http://dribbble.com/shots/889189-Gradients-Time/attachments/96560

Tip of the hat to @sheelc and @vito for basically writing all of this while
I explained how to use my bizarro version of emacs and complained that the
implementation wasn't performant enough and that window.setInterval is so
much superior to a recursive window.setTimeout.

Turns out I was wrong, OS X tends to reduce performance of applications
when running on a low battery, so it just *seemed* to be worse than how
we used setInterval before. My apologies, @sheelc.