Fade in / out waveform based on volume across the spectrum
This is just a simple linear filter, and I imagine we might want to
scale this as a logarithm or maybe just quadratically, depending
on how noticable this is during the day. I'd like it if the
CLAP during standup and pingpong was noticable, but the dull roar of
the office shouldn't trigger it.